Revert "feat: #20 share flow redesign with visual snapshot export (#726)"
This reverts commit 349b2e5404a75e773958c76e7b4a43e6c4a9c00c.
