+++ package.json
@@ -56,7 +56,6 @@
     "fflate": "^0.8.2",
     "geotiff": "^3.0.5",
     "jose": "^6.2.1",
-    "jspdf": "^4.2.1",
     "lucide-react": "^1.7.0",
     "maplibre-gl": "^5.20.0",
     "react": "^19.2.0",
@@ -65,7 +64,6 @@
     "react-markdown": "^9.1.0",
     "remark-gfm": "^4.0.1",
     "simple-icons": "^16.14.0",
-    "svg2pdf.js": "^2.7.0",
     "zod": "^4.3.6",
     "zustand": "^5.0.11"
   },
