+++ src/index.css
@@ -5036,139 +5036,3 @@ html.panorama-gesture-lock body {
   max-height: none;
   overflow: visible;
 }
-
-/* =========================================================
-   Share Modal — tab shell
-   ========================================================= */
-
-.share-modal {
-  width: min(760px, 96vw);
-}
-
-.share-modal-tabs {
-  display: flex;
-  gap: 2px;
-  border-bottom: 1px solid var(--border);
-  padding-bottom: 0;
-  margin-bottom: 4px;
-}
-
-.share-modal-tab {
-  padding: 6px 14px;
-  border: none;
-  background: none;
-  cursor: pointer;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: var(--text-muted);
-  border-bottom: 2px solid transparent;
-  margin-bottom: -1px;
-  border-radius: 0;
-  transition: color 0.15s, border-color 0.15s;
-}
-
-.share-modal-tab:hover {
-  color: var(--text);
-}
-
-.share-modal-tab.is-active {
-  color: var(--accent);
-  border-bottom-color: var(--accent);
-}
-
-/* =========================================================
-   Export Composer
-   ========================================================= */
-
-.export-composer {
-  display: grid;
-  gap: 14px;
-  padding-top: 6px;
-}
-
-.export-preview-wrap {
-  position: relative;
-  background: var(--surface-2);
-  border: 1px solid var(--border);
-  border-radius: var(--radius-card);
-  overflow: hidden;
-  min-height: 120px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.export-preview-canvas {
-  display: block;
-  width: 100%;
-  border-radius: var(--radius-card);
-}
-
-.export-preview-canvas.is-loading {
-  opacity: 0.5;
-}
-
-.export-preview-status {
-  position: absolute;
-  bottom: 8px;
-  left: 50%;
-  transform: translateX(-50%);
-  white-space: nowrap;
-  pointer-events: none;
-}
-
-.export-options {
-  display: grid;
-  gap: 10px;
-}
-
-.export-fieldset {
-  border: 1px solid var(--border);
-  border-radius: var(--radius-card);
-  padding: 10px 12px;
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.export-legend {
-  font-size: 0.75rem;
-  font-weight: 600;
-  color: var(--text-muted);
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  padding: 0 4px;
-}
-
-.export-toggle-label {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 0.875rem;
-  cursor: pointer;
-}
-
-.export-toggle-label.is-disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.export-row {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.export-label {
-  font-size: 0.875rem;
-  font-weight: 500;
-  min-width: 60px;
-  flex-shrink: 0;
-}
-
-.export-actions {
-  display: flex;
-  gap: 8px;
-  flex-wrap: wrap;
-}
